requirements.txt
accelerate
fastapi
httpx
langchain-openai
neo4j
numpy
openai
pandas
python-dotenv
scikit-learn
torch
transformers
tqdm
uvicorn[standard]
